[Rouge] PML Nodes
Now the lexer is able to highlight any valid node and its closing `]`
delimiter. If a node with an unknown tag is encountered, it's marked
as `Error`, along with its closing delimiter.

Although the lexer currently treats all nodes alike (i.e. no special
handling of their contents, e.g. like formatting in bold or italic if
it's a font node), it's able to track nodes' nesting on the stack, and
pop correctly whenever a node's closing `]` is encountered.

The code is structured to allow customizing node behaviours according
to tag, without having to rewrite the current code but just adding to
it additional checks.