Highlighters Schemes
Create new `syntax-hl/schemes/` folder and add to it the ColorImpact 4
source project with all the colour schemes, plus their exported swatches
preview images and colour specs sheets in plaintext.

Move the credits info for the Gogh Flat scheme from the README of the
`rouge/` folder to that of the new `schemes/` folder, and provide a link
to the new document instead (since the same schemes will be used with
different highlighters, let's avoid credits redundancy).