[Rouge] Update Docs
Update the AsciiDoc Examples document with a new section covering the
extra styles included in the docinfo file, along with instructions on
how to use it in custom projects.

Move custom attributes to a new `_attr-custom.adoc` shared file and
tweak the `Rakefile` to track it as a requirement for building the
HTML docs in the `rouge` task.