[mustache] New AsciiDoc Template
Add to `mustache/` a new `tags-ref__asciidoc.mustache` template which
yields the `tags-ref.asciidoc` source document, which the Rakefile
automatically converts to `tags-ref.html`.

This document is the AsciiDoc version of the pre-existing `tags-info.md`
doc, but it also contains tagged region comments which will allow part
of its contents to be included in the new PML Syntax Guide, so that the
PML nodes/tags info parts can be automated via Mustache.