Rake Workaround
Tweak `pandoc` task in `Rakefile` to skip converting
to JSON the `elements/math.markdown` test which always
fails the JSON prettify step, resulting in either:

* An empty JSON file (which fails EClint validation).
* A Rake crash.

The problem seems to be related to the complexity
of Tex Math representation in the pandoc AST when
converted to JSON, which causes problems in the JSON
prettify stage. The easiest fix in this case was just
skipping that problematic file altogether, since we
don't really need its JSON representation for study
purposes.