[CSS] New Default Stylesheet
The new `css__default-new/` folder contains
a WIP attempt to port the default PML 3.1.0
stylesheets to Sass, splitting them into
SMACSS rules modules.

Currently the original CSS was just spread
across the various Sass sources and modules,
unchanged, and normalize.css v8.0.1 was added
as a Sass module (with minor tweaks).

This will serve as the starting point for a
new Sass based and SMACSS oriented version
of the default stylesheets, which could
eventually become the official version.