[Pandoc Writer] Migrate to PML 4 & pandoc 3
The new pandoc 3 release introduced breaking changes
to the Lua filters API. This commit patches the custom
writer by applying a quick workaround to allow using it
without having to refactor it.

Fixes are also applied to match the breaking changes
in the syntax of PML 4, to ensure that all generated
PML documents will pass conversion.

README documents are updated accordingly, with all
references to PML, pandoc and Lua version numbers
being updated to PML 4, pandoc 3 and Lua 5.4.

Links to pandoc and Lua documentation are updated
to match the current versions used by pandoc.